Fix saving text targets in chains
Text targets can have dynamic text, i.e. text not loaded from
an existing test file but rather passed as a string. Since this
is a supported scenario when loading a chain of steps, it's
naturally expected that the reverse is also possible and that
the chain won't error out when we try to dump it.

Also add tests for this scenario since we didn't have any for
chain targets, and fix errors revealed by the tests in the way
the steps file was generated.