Add support for scoped path resolution
The current path resolver uses a shared list of paths, and this
might lead to conflicts when searching for files as the caller
has no control over which paths are added to the shared list.

With a scoped path we can have a controlled context inside which
only the added paths will be searched, adding the ability to use
shorter names without the fear of conflicts.